requirements.txt
pandas
openpyxl
Faker
python-decouple
python-dotenv
SQLAlchemy
